taxonomy: Add “concentrate” as distinct ingredient (#12031)
There are currently 288 products with `en:concentrate` in their
ingredients[1], compared with 45 with `en:concentrated extract`[2].
That’s “concentrate” being used more than 5 times more than the
“extract” term, but only the “extract” variant is in
`food/ingredients.txt`, leading to all the others to go unrecognised.

This adds “concentrate” as its own, distinct ingredient. Hopefully we’ll
be able to parse it better in the future.

See also
https://openfoodfacts.slack.com/archives/C06A7LENM/p1749983011998189
and comments on
https://github.com/openfoodfacts/openfoodfacts-server/pull/12031

[1]
https://world.openfoodfacts.org/facets/ingredients/concentrate
[2]
https://world.openfoodfacts.org/facets/ingredients/concentrated-extract