feat: add Nutri-Score % of fruits / vegetables / nuts / legumes from GS1 (#12005)
GS1 has added support for some fields that are necessary to compute the
Nutri-Score, like adding the % of fruits / vegetables / legumes. We do
get some of those from some producers (e.g Bonduelle, but for the old
Nutri-Score fields it seems).

This PR adds some support + some comments and links in the code.

There are more fields we could add (which category should be used for
the Nutri-Score: fats, nuts etc.). But so far I have not seen it in the
data we receive yet.

---------

Co-authored-by: hangy <hangy@hangy.de>
Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>