feat: add no added sugars label from gs1 (#12084)
GS1 provides several claims for its products that are not yet imported.
In this PR the claim "no added sugars" is now imported from the GS1
provided files into the products _labels_ field. New GS1 files were also
added to test that the new claim is correctly imported.

This PR is related to the issue #12006.