fix: make robots.txt static (#11961)
robots.txt is currently generated for every request by Product Opener,
so that we can have different rules:
- based on user-agent (e.g. to allow some facets like /brands for some
crawlers)
- based on cc + lc (to be able to list facets like /facets/marques in
French for instance, and also to disallow crawling of subdomaines like
fr-en)

It's not very convenient though, as some robots use tons of ips, and
some of those send thousands of requests for robots.txt, that go to our
low priority Apache, which is often overloaded and not responding.

I would really like to make robots.txt static again, so that it can be
served instantly by nginx.

This PR contains a simple robots.txt:

- it disallows /facets for everyone, except Googlebog and Bingbot
- it disallows / for a bunch of robots that have been crawling us hard
in the past

It doesn't solve the issues of disallowing things like fr-en and
/facets/dates-d-ajout, or sending slightly different robots.txt to some
crawlers.

We could pre-generate a bunch of files like /html/robots/[cc]-[lc].txt
for all combinations of countries and languages.

We could also normalize URLs to English, so that /facets/marque becomes
/facet/brand as we discussed when we added the /facets/ prefix