taxonomy: Slight overhaul of vegan classification
A number of ingredients that are on their surface vegan actually undergo
some kind of treatment or processing using animal products, which would
make them not vegan. In many cases, this would get reflected in
ingredient lists, like gelatin in jams and candy etc., but in some cases
this is done in ways that are not going to be reflected in an ingredient
list, e.g., because there won’t be one—like for fruits.

This also means that just because something is “vegetable”, “extracted
from plants”, “plant based”, etc. it doesn’t automatically make them
vegan, so those labels have been moved to be synonyms (or children) of
vegetarian rather than vegan.

This approach is the same that is already used in `beauty/labels.txt`
where “no animal derived ingredients” implying vegetarian status rather
than vegan status:
```
< en:Vegetarian
en: No animal derived ingredients
```

Specific changes
================

Labels
------

* “Plant based” has been moved to sit between Vegetarian and Vegan, and
  relevant synonyms that were previously attached to `en:vegan` have (to
  the best of my knowledge) been moved to that.
* A lot of labels have gotten `incompatible_with` tags set for
  vegetarian or vegan, as applicable.
* A lot of labels have had `en:Non-vegan` or `en:Non-vegetarian` set as
  parents, to take advantage of the code in `Ingredients.pm` that sets
  non-vegan/-vegetarian status on products based on these labels.
  * This will hopefully properly mark some products as non-vegan that
    might currently be “maybe” or “unknown” due to missing or unknown
    ingredients.

Ingredients
-----------

* `en:animal` is set to non-vegan, as anything that is derived from a
  non-human animal is, by definition, not vegan.
* Vitamins have been changed from `vegan:en: yes` to `maybe`, as some
  vitamins may be extracted or otherwise derived from non-human animals.
* Set `vegan` and/or `vegetarian` attributes for some that weren’t
  recognised as such via parents or otherwise.

Categories
----------

* A new `en: Vegetarian products` category was added.
* A lot of categories had their ancestry fixed/updated to include
  `en: Veg(egetari|)an` where appropriate.
* A lot of categories had `incompatible_with:en:` set to either the
  “Vegetarian products” or “Vegan products” category.
* A few products were cleaned up with wrong(/old?) claims of being
  vegan or vegetarian.