docs: a few fixes in the comments for Nutriscore (#12171)
A few fixes in the comments.

Some remarks. I do not know what is the use of the
`[nutrient]_points_max` property, so the comment contains question
marks.

In the 2023 version, I have added a comment for nutrient `salt` but I
have not removed `sodium`. The provisional comment for `salt` contains
only question marks, please write something interesting.

There is a problem with the signatures of functions
`get_value_with_one_less_negative_point_2023` and
`get_value_with_one_more_positive_point_2023`, but it is given in a
separate issue.

---------

Co-authored-by: Stéphane Gigandet <stephane@openfoodfacts.org>