feat: update new nutrition schema (#12239)
In this PR the new nutrition schema is updated based on the changes on
[2025 - API - Nutrition Facts Schema and
API](https://docs.google.com/document/d/19ZRrlWJraJm61E6U7AwxQ1uubPDvmSuNfl9F1oLC0Tg).
The main changes are new names for the nutrient sets, simplification of
the generated set object and quality check field addition.

The OpenAPI documentation and the new features based on the new schema
are updated in consequence.