taxonomy: correction coating ingredients (#12432)
Correction of coating ingredients. "en: berry coating" was in berry,
which would count it as a fruit for the nutriscore. However, berry
coating is susceptible to contain other ingredients, such as sugar, and
should'nt be classified as a fruit. Same for "chocolate and hazelnut
coating" which was indirectly in "hazelnut" through "hazelnut coating".
It classifies it as a "fruit, vegetable or nut" for the nutriscore
calculation.