fix: default sort by last modified date on OPF (#12479)
This PR makes OPF / OBF / OPFF have a default sort on last modification
date instead of popularity (which is not very relevant for products and
makes things like cigarettes come up first on OPF because they are
scanned a lot with the app).

Also removed an old sort key last_modified_t_complete_first that is not
generated anymore.

---------

Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>