fix: check product type for some GS1 prefix checks (#12533)
### What

Adds additional checks to the GS1 prefix QA checks for whether an item
is already properly classified as a `product` or as a `beauty` item
before adding a QA warning/notice to them.

A warning about an item being a book/periodical or cosmetic item and a
note to move them to OPF or OBF when that’s where they already are, is
misleading.

This _does not_ add additional texts/warnings to distinguish e.g. books
entered on OBF to be “not food”, leaving that change for someone else to
do.

### Related issue(s) and discussion

- Fixes
https://github.com/openfoodfacts/openfoodfacts-server/issues/12532

---------

Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>
Co-authored-by: Stéphane Gigandet <stephane@openfoodfacts.org>