ci: add additional feedback when PR action is done (#12541)
Right now there’s feedback (in the form of a reaction) when a PR action
is seen (and thus begun being acted on), but there’s no feedback when
it’s done being processed, esp. when there’s no resulting changes and
thus no commits.

This replaces the 👀 reaction with 🚀 at the end of the action, letting
the caller know the action has finished running.

(It also utilises YAML anchors to not have to repeat code.)