feat: WIP - Support for new nutrition schema: product edit, API v2 + v3, import and export (#12475)
This PR is a work in progress to implement support for the new nutrition
schema:
https://docs.google.com/document/d/19ZRrlWJraJm61E6U7AwxQ1uubPDvmSuNfl9F1oLC0Tg/edit?tab=t.0

Included changes:
- new product edit form
- API v2 old nutrition fields backward compatibility
- API v2 new nutrition fields for new nutrition schema
- API v3 new nutrition fields for new nutrition schema
- export_csv_file support
- import_csv_file support

This PR changes the product edit form so that it can edit the different
possible nutrition input sets for 1 specific source (the "packaging"
source on the public platform, or the "manufacturer" (or USDA etc.)
source on the producers platform).

Checkboxes are displayed to show the different possible input set
columns (some of them are hidden if we don't have data yet).

In theory the units can be different for the same nutrient in different
input sets. In practice it should not be often the case, so unless we do
have existing different units, only 1 global unit column is shown.

<img width="490" height="765" alt="image"
src="https://github.com/user-attachments/assets/850b88b3-d2ce-4da9-b2ed-6c87d9f4f81a"
/>

Otherwise we can show units by input set:

<img width="547" height="765" alt="image"
src="https://github.com/user-attachments/assets/42cb81ea-c016-4752-a449-c07d083de0e0"
/>

Compared to the current product edit form, we do not support unknown
nutrients, we limit users to nutrients that exist in the taxonomy.

So the system to add lines with labels that can be typed in or selected
from suggestions has been changed: there is now a button at the bottom
to display additional nutrients. This has the benefit of greatly
simplifying the code.

Also removed: the salt and sodium rows are now independent, we don't
compute one from the other anymore.

To be fixed:
- nutrient checks (e.g. sugars <= carbohydrates) should be made to work
on all columns

The new product edit form still uses the API v2 parameters. So I added
new fields like
nutrition_input_sets_${preparation}_${per}_${nid}_value_string to pass
values.

The same function will be used to add support for those fields in the
actual API v2, so that existing apps can set values for the input sets,
without switching to the API v3 (which does not handle nutrients yet).

API v2 will continue to support existing nutrition fields and "guess" in
which input set they belong.