ci: remove `off-org` from deploy matrix (#12663)
Right now, every commit to `main` as well as every `v*.*.*` release will
result in a red “x”/failed status, as the deploy workflow exits early
for deploying to off.org.

This removes `off-org` from the build matrix entirely, to not run this.
Hopefully this will lead us to not see all those red x’s all the time,
only when there’s something _actually_ wrong. :)

Note that this _doesn’t_ simplify the workflow based on there only being
one entry in the matrix, since this allows for a future easy reversion
of this commit or otherwise re-insert `off-org` into the matrix.

Co-authored-by: Alex Garel <alex@openfoodfacts.org>