fix: selected country now shows on country filter dropdown (#12720)
This PR fixes the error where after selecting a country filter on the
website, the results are not updated correctly. Upon being navigated to
the selected country, users are shows the default country dropdown with
no indication of what country's OPF server they have been redirected to.

This makes it unclear to the user which country filter is currently
active, so this PR shows the current country selected on the new
redirected server.

Fixes #12254